Extend cluster autoscaler to check AUTOSCALER_ENV_VARS in kube-env
This is the second part of the fix in https://github.com/kubernetes/kubernetes/pull/61119

This provides a temporary way for the cluster autoscaler to get at
values that were removed from kube-env in https://github.com/kubernetes/kubernetes/pull/60020.
Ideally this information will eventually be available via e.g. the Cluster API,
because kube-env is an internal interface that carries no stability
guarantees.